Final_RAG_Agent/RAG_AI_Agent/requirements.txt
langchain
langchain-openai
langchain_community
langgraph
pypdf
pydantic
python-dotenv
python-docx
faiss-cpu
docx2txt